Added GPL license to project
The reason that I chose to use the GPL license instead of a more permissive license is simple. I don't want the free software created by me and future contributors to be used in a closed-source (presumably commercial) product. The primary benefactor of this software should always be the users and the community.